refactor: moved GoogleResult into main body of Model package
