+++ src/test/java/org/kunekune/PiglioTech/model/GoogleResultTest.java
@@ -2,7 +2,6 @@ package org.kunekune.PiglioTech.model;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.kunekune.PiglioTech.model.google.GoogleResult;
 
 import static org.junit.jupiter.api.Assertions.*;
